+++ printf.c
@@ -4,7 +4,7 @@
 
 int main(void)
 {
-    float x = 3.10;
+    double x = 3.10;
     int i = 42;
 
     printf("i=%d\n", i);
